refactor(calendar): address CodeRabbit review on assignment-map PR
- geofence-coords.service: move GeofencePoint/GeofenceCoordsResponse to a
  colocated *.types.ts; trim+normalize origin/destination before the SWR key
  (avoids whitespace-only fetches / duplicate cache entries).
- geofence-coords route: use @/ path aliases instead of ../../ relatives.
- assignment-form: explicit `eta != null` checks instead of truthiness.

Skipped CodeRabbit's "typed pgrest error class" nit — every fetch in
pgrest-client throws a generic Error; a one-off custom error would be
inconsistent with the file.

Refs #714
