+++ turbo-repo/apps/app/src/features/calendar/services/geofence-coords.service.types.ts
@@ -1,0 +1,9 @@
+export interface GeofencePoint {
+  latitude: number;
+  longitude: number;
+}
+
+export interface GeofenceCoordsResponse {
+  origin: GeofencePoint | null;
+  destination: GeofencePoint | null;
+}
